internal_docs_qa/requirements.txt
langchain>=0.2
langchain-community>=0.2
sentence-transformers>=2.2
faiss-cpu>=1.7
python-dotenv>=1.0
unstructured
transformers>=4.40
torch
gradio>=4.0